chore: 메인페이지 카테고리 섹션 컴포넌트 분할
