+++ src/pages/Main.css
@@ -15,25 +15,6 @@
     padding: 0 20px;
 }
 
-.categories {
-    display: grid;
-    grid-template-columns: repeat(6, 1fr); /* 기본적으로는 2행 6열 */
-    justify-content: center;
-    gap: 40px;
-    padding: 40px calc(8% + 120px);
-}
-
-.categoryContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-.categoryText {
-    margin-top: 10px;
-}
-
 .buttonContainer {
     display: flex;
     justify-content: center;
@@ -68,9 +49,3 @@
         padding: 40px 20px;
     }
 }
-
-@media (max-width: 1200px) {
-    .categories {
-      grid-template-columns: repeat(4, 1fr); /* 1200px 이하에서는 4행 4열 */
-    }
-  }
